// apps/web-app/src/server/scheduling/types/type.ts
export type Room = "EN100" | "EN200" | "EN300" | "EN400" | "EN500" | "EN600";
export type DaysOfWeek = "M" | "T" | "W" | "TH" | "F" | "S";
export type Time =
  | "0700"
  | "0730"
  | "0800"
  | "0830"
  | "0900"
  | "0930"
  | "1000"
  | "1030"
  | "1100"
  | "1130"
  | "1200"
  | "1230"
  | "1300"
  | "1330"
  | "1400"
  | "1430"
  | "1500"
  | "1530"
  | "1600"
  | "1630"
  | "1700"
  | "1730"
  | "1800"
  | "1830"
  | "1900"
  | "1930"
  | "2000"
  | "2030"
  | "2100"
  | "2130";